moved key matching  to validation
